refactor : 엔티티 경로 변경 - #1
