+++ src/main/java/kr/inuappcenterportal/inuportal/controller/ReportController.java
@@ -10,7 +10,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
-import kr.inuappcenterportal.inuportal.domain.Member;
+import kr.inuappcenterportal.inuportal.domain.member.Member;
 import kr.inuappcenterportal.inuportal.dto.ReportListResponseDto;
 import kr.inuappcenterportal.inuportal.dto.ReportRequestDto;
 import kr.inuappcenterportal.inuportal.dto.ResponseDto;
