+++ src/main/java/kr/inuappcenterportal/inuportal/domain/replylike/ReplyLike.java
@@ -1,6 +1,8 @@
-package kr.inuappcenterportal.inuportal.domain;
+package kr.inuappcenterportal.inuportal.domain.replylike;
 
 import jakarta.persistence.*;
+import kr.inuappcenterportal.inuportal.domain.member.Member;
+import kr.inuappcenterportal.inuportal.domain.reply.Reply;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
