+++ src/main/java/kr/inuappcenterportal/inuportal/domain/schedule/Schedule.java
@@ -1,4 +1,4 @@
-package kr.inuappcenterportal.inuportal.domain;
+package kr.inuappcenterportal.inuportal.domain.schedule;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
