+++ src/main/java/kr/inuappcenterportal/inuportal/dto/NoticeListResponseDto.java
@@ -1,7 +1,7 @@
 package kr.inuappcenterportal.inuportal.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import kr.inuappcenterportal.inuportal.domain.Notice;
+import kr.inuappcenterportal.inuportal.domain.notice.Notice;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
