+++ src/main/java/kr/inuappcenterportal/inuportal/dto/ScheduleResponseDto.java
@@ -1,7 +1,7 @@
 package kr.inuappcenterportal.inuportal.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import kr.inuappcenterportal.inuportal.domain.Schedule;
+import kr.inuappcenterportal.inuportal.domain.schedule.Schedule;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
