+++ src/main/java/kr/inuappcenterportal/inuportal/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package kr.inuappcenterportal.inuportal.repository;
 
-import kr.inuappcenterportal.inuportal.domain.Category;
+import kr.inuappcenterportal.inuportal.domain.category.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
