+++ src/main/java/kr/inuappcenterportal/inuportal/repository/FireRepository.java
@@ -1,6 +1,6 @@
 package kr.inuappcenterportal.inuportal.repository;
 
-import kr.inuappcenterportal.inuportal.domain.Fire;
+import kr.inuappcenterportal.inuportal.domain.fire.Fire;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
