+++ src/main/java/kr/inuappcenterportal/inuportal/repository/FolderRepository.java
@@ -1,7 +1,7 @@
 package kr.inuappcenterportal.inuportal.repository;
 
-import kr.inuappcenterportal.inuportal.domain.Folder;
-import kr.inuappcenterportal.inuportal.domain.Member;
+import kr.inuappcenterportal.inuportal.domain.folder.Folder;
+import kr.inuappcenterportal.inuportal.domain.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
