+++ src/main/java/kr/inuappcenterportal/inuportal/repository/NoticeRepository.java
@@ -1,6 +1,6 @@
 package kr.inuappcenterportal.inuportal.repository;
 
-import kr.inuappcenterportal.inuportal.domain.Notice;
+import kr.inuappcenterportal.inuportal.domain.notice.Notice;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
