+++ src/test/java/kr/inuappcenterportal/inuportal/controller/MemberControllerTest.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import kr.inuappcenterportal.inuportal.config.SecurityConfig;
 import kr.inuappcenterportal.inuportal.config.TokenProvider;
-import kr.inuappcenterportal.inuportal.domain.Member;
+import kr.inuappcenterportal.inuportal.domain.member.Member;
 import kr.inuappcenterportal.inuportal.dto.LoginDto;
 import kr.inuappcenterportal.inuportal.dto.MemberResponseDto;
 import kr.inuappcenterportal.inuportal.dto.MemberUpdateNicknameDto;
