+++ src/test/java/kr/inuappcenterportal/inuportal/controller/ReportControllerTest.java
@@ -1,12 +1,11 @@
 package kr.inuappcenterportal.inuportal.controller;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import kr.inuappcenterportal.inuportal.config.SecurityConfig;
 import kr.inuappcenterportal.inuportal.config.TokenProvider;
-import kr.inuappcenterportal.inuportal.domain.Member;
-import kr.inuappcenterportal.inuportal.domain.Report;
+import kr.inuappcenterportal.inuportal.domain.member.Member;
+import kr.inuappcenterportal.inuportal.domain.report.Report;
 import kr.inuappcenterportal.inuportal.dto.ReportListResponseDto;
 import kr.inuappcenterportal.inuportal.dto.ReportRequestDto;
 import kr.inuappcenterportal.inuportal.service.ReportService;
@@ -24,7 +23,6 @@ import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
 import org.springframework.http.MediaType;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 
 import java.lang.reflect.Field;
