+++ src/test/java/kr/inuappcenterportal/inuportal/service/MemberServiceTest.java
@@ -2,7 +2,7 @@ package kr.inuappcenterportal.inuportal.service;
 
 
 import kr.inuappcenterportal.inuportal.config.TokenProvider;
-import kr.inuappcenterportal.inuportal.domain.Member;
+import kr.inuappcenterportal.inuportal.domain.member.Member;
 import kr.inuappcenterportal.inuportal.oracleRepository.SchoolLoginRepository;
 import kr.inuappcenterportal.inuportal.repository.MemberRepository;
 import org.junit.jupiter.api.BeforeEach;
